Major rewrite of the frontend, add support for v2 pkg format
Rewrite the frontend significantly and expand on the instructions to
guide the user through the whole installation process, including the
non-bootloader steps like extracting a Rockbox build to the SD card.

Also adds support for the v2 package format (using bootloader2.ucl)
which brings this up to date with the latest bootloader changes.